requirements.txt
aiogram
numpy
pandas
pre-commit
pytest
pytest-json-report
python-dotenv
